pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "incident-response-env"
version = "1.0.0"
description = "DevOps Incident Response OpenEnv Environment"
requires-python = ">=3.11"
dependencies = [
    "openenv-core>=0.1.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.29.0",
    "pydantic>=2.0.0",
]

[tool.pytest.ini_options]
testpaths = ["."]
python_files = ["test_*.py"]